Module ra_hir_def::item_tree

A simplified AST that only contains items.

Structs

Const

Enum

Expr

ExternCrate

Field
A single field of an enum variant or struct

FileItemTreeId

Function

GenericParamsId

IdRange

Impl

Import
A desugared use import.

ItemTree
The item tree of a source file.

MacroCall

Mod

RawVisibilityId

Static

Struct

Trait

TypeAlias

Union

Variant

Enums

AssocItem

AttrOwner

Fields

ModItem

ModKind

StructDefKind

Traits

ItemTreeNode
Trait implemented by all item nodes in the item tree.

Type Definitions

ItemTreeId