Crate ra_hir_def

hir_def crate contains everything between macro expansion and type inference.

It defines various items (structs, enums, traits) which comprises Rust code, as well as an algorithm for resolving paths to such entities.

Note that hir_def is a work in progress, so not all of the above is actually true.

Modules

adt
Defines hir-level representation of structs, enums and unions

attr
A higher level attributes based on TokenTree, with also some shortcuts.

builtin_type
This module defines built-in types.

db
Defines database & queries for name resolution.

diagnostics
Diagnostics produced by hir_def.

nameres
This module implements import-resolution/macro expansion algorithm.

path
FIXME: write short doc here

type_ref
HIR for references to types. Paths in these are not yet resolved. They can be directly created from an ast::TypeRef, without further queries.

Structs

ConstId

CrateModuleId
An ID of a module, local to a specific crate

EnumId

EnumVariantId

FunctionId

ItemLoc

LocalEnumVariantId

LocalStructFieldId

LocationCtx

ModuleId

StaticId

StructFieldId

StructId

TraitId

TypeAliasId

UnionId

Enums

AdtId
A Data Type

ModuleDefId
The defs which can be visible in the module.

ModuleSource

VariantId

Traits

AstItemDef